Refactor stream endpoint, fix encoder thread safety and improve tests
